Fixed app/client handlers on Windows. The calling convention for CEF callbacks needs to be __stdcall (cef_export.h).
The CEF_CALLBACK was misplaced in cef_app/cef_client header
files.

+++ handlers/cef_base.h
@@ -63,7 +63,7 @@ void initialize_cef_base(cef_base_t* base) {
     // Let's print the size in case sizeof was used
     // on a pointer instead of a structure. In such
     // case the number will be very high.
-    printf("cef_base_t.size = %lu\n", size);
+    printf("cef_base_t.size = %lu\n", (unsigned long)size);
     if (size <= 0) {
         printf("FATAL: initialize_cef_base failed, size member not set\n");
         _exit(1);
